chore: chagned uint16 to uint64 for bandwidthBytesPerSecond

+++ src/types/SharedTypes.sol
@@ -38,7 +38,7 @@ library SharedTypes {
         /// @dev Valid range: 0-10000 (basis points, e.g. 7550 = 75.50%). 0 means "don't care".
         uint16 retrievabilityBps;
         /// @dev Capped at ~64 Gbps
-        uint16 bandwidthBytesPerSecond;
+        uint64 bandwidthBytesPerSecond;
         uint16 latencyMs;
         /// @dev Valid range: 0-100. 0 means "don't care".
         uint8 indexingPct;
